(PUP-6220) Move PObjectType resolution behavior to abstract super class
This commit refactors the PObjectClass into two classes. One abstract
super class that deals with the general aspects of deferred resolution
of content, and one concrete subclass that inherits that behavior.
