+++ lib/puppet/pops/types/p_object_type.rb
@@ -4,7 +4,7 @@ module Puppet::Pops
 module Types
 
 # @api public
-class PObjectType < PAnyType
+class PObjectType < PMetaType
   KEY_ANNOTATIONS = 'annotations'.freeze
   KEY_ATTRIBUTES = 'attributes'.freeze
   KEY_CHECKS = 'checks'.freeze
@@ -13,7 +13,6 @@ class PObjectType < PAnyType
   KEY_FINAL = 'final'.freeze
   KEY_FUNCTIONS = 'functions'.freeze
   KEY_KIND = 'kind'.freeze
-  KEY_NAME = 'name'.freeze
   KEY_OVERRIDE = 'override'.freeze
   KEY_PARENT = 'parent'.freeze
   KEY_TYPE = 'type'.freeze
@@ -452,58 +451,6 @@ class PObjectType < PAnyType
   # @api private
   def include_class_in_equality?
     @equality_include_type && !(@parent.is_a?(PObjectType) && parent.include_class_in_equality?)
-  end
-
-  # Called from the TypeParser once it has found a type using the Loader. The TypeParser will
-  # interpret the contained expression and the resolved type is remembered. This method also
-  # checks and remembers if the resolve type contains self recursion.
-  #
-  # @param type_parser [TypeParser] type parser that will interpret the type expression
-  # @param loader [Loader::Loader] loader to use when loading type aliases
-  # @return [PObjectType] the receiver of the call, i.e. `self`
-  # @api private
-  def resolve(type_parser, loader)
-    unless @i12n_hash_expression.nil?
-      @self_recursion = true # assumed while it being found out below
-
-      i12n_hash_expression = @i12n_hash_expression
-      @i12n_hash_expression = nil
-      if i12n_hash_expression.is_a?(Model::LiteralHash)
-        i12n_hash = resolve_literal_hash(type_parser, loader, i12n_hash_expression)
-      else
-        i12n_hash = resolve_hash(type_parser, loader, i12n_hash_expression)
-      end
-      initialize_from_hash(i12n_hash)
-
-      # Find out if this type is recursive. A recursive type has performance implications
-      # on several methods and this knowledge is used to avoid that for non-recursive
-      # types.
-      guard = RecursionGuard.new
-      accept(NoopTypeAcceptor::INSTANCE, guard)
-      @self_recursion = guard.recursive_this?(self)
-    end
-    self
-  end
-
-  def resolve_literal_hash(type_parser, loader, i12n_hash_expression)
-    type_parser.interpret_LiteralHash(i12n_hash_expression, loader)
-  end
-
-  def resolve_hash(type_parser, loader, i12n_hash)
-    resolve_type_refs(type_parser, loader, i12n_hash)
-  end
-
-  def resolve_type_refs(type_parser, loader, o)
-    case o
-    when Hash
-      Hash[o.map { |k, v| [resolve_type_refs(type_parser, loader, k), resolve_type_refs(type_parser, loader, v)] }]
-    when Array
-      o.map { |e| resolve_type_refs(type_parser, loader, e) }
-    when PTypeReferenceType
-      o.resolve(type_parser, loader)
-    else
-      o
-    end
   end
 
   # @api private
