+++ lib/puppet/pops/types/types.rb
@@ -2848,6 +2848,7 @@ end
 require 'puppet/pops/pcore'
 
 require_relative 'puppet_object'
+require_relative 'p_meta_type'
 require_relative 'p_object_type'
 require_relative 'p_runtime_type'
 require_relative 'p_sem_ver_type'
